requirements.txt
# AI Providers
groq==0.13.0
google-generativeai==0.8.3

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Utilities
python-dotenv==1.0.1
colorama==0.4.6
rich==13.9.4

# Logging
loguru==0.7.3